refactoring: moving files to the right folders
